+++ src/main/java/io/github/astrapi69/swing/robot/RobotExtensions.java
@@ -56,7 +56,6 @@ public class RobotExtensions
 
 	/**
 	 * Clicks the left mouse button
-	 * 
 	 * @param robot
 	 *            the robot
 	 * @param x
@@ -64,14 +63,12 @@ public class RobotExtensions
 	 * @param y
 	 *            the Y position
 	 */
-	public static void clickLeftMouseButton(Robot robot, int x, int y)
-	{
+	public static void clickLeftMouseButton(Robot robot, int x, int y) {
 		clickMouseButton(robot, x, y, InputEvent.BUTTON1_DOWN_MASK);
 	}
 
 	/**
 	 * Clicks the middle mouse button
-	 * 
 	 * @param robot
 	 *            the robot
 	 * @param x
@@ -79,14 +76,12 @@ public class RobotExtensions
 	 * @param y
 	 *            the Y position
 	 */
-	public static void clickMiddleMouseButton(Robot robot, int x, int y)
-	{
+	public static void clickMiddleMouseButton(Robot robot, int x, int y) {
 		clickMouseButton(robot, x, y, InputEvent.BUTTON2_DOWN_MASK);
 	}
 
 	/**
 	 * Clicks the right mouse button
-	 * 
 	 * @param robot
 	 *            the robot
 	 * @param x
@@ -94,8 +89,7 @@ public class RobotExtensions
 	 * @param y
 	 *            the Y position
 	 */
-	public static void clickRightMouseButton(Robot robot, int x, int y)
-	{
+	public static void clickRightMouseButton(Robot robot, int x, int y) {
 		clickMouseButton(robot, x, y, InputEvent.BUTTON3_DOWN_MASK);
 	}